Particles Work
Finally, some basic form of particles work.

+++ src/psystem/spawnInfo.h
@@ -55,13 +55,13 @@ struct PSpawnInfo
     float minAngularSpeed; // Minimum angular speed.
     float maxAngularSpeed; // Maximum angular speed.
     float spawnTime; // How many seconds the system can spawn particles for. 0 for infinite.
-    unsigned int scaleRandomness; // Varies random scale of particles.
+    float scaleRandomness; // Varies random scale of particles.
     float lifetime; // How many frames particles exist for.
     float lifetimeRandomness; // Random lifetime of particles.
-    unsigned int speedRandomness; // Random speed of particles.
+    float speedRandomness; // Random speed of particles.
     float spawnPeriod; // Seconds between spawning frames.
-    unsigned int alpha; // How opaque particles are.
-    unsigned int speedFalloff; // How fast speed decelerates or accelerates.
+    float alpha; // How opaque particles are.
+    float speedFalloff; // How fast speed decelerates or accelerates.
     unsigned int altLength; // Alternate number of frames to use for transitions instead of lifetime.
     float velStretchFactor; // How much to stretch by velocity.
     unsigned int logTexRepeatHorizontal; // Repeat texture in X direction count.
